feat(queue data item): replace placeholders in SQL upsertDataItem PE-6220
