+++ src/types.d.ts
@@ -529,16 +529,3 @@ export type KVBufferStore = {
   del(key: string): Promise<void>;
   has(key: string): Promise<boolean>;
 };
-
-/** Accepted in queue data item route fields as normalized b64 */
-export interface QueueDataItemHeaders {
-  content_type: string;
-  data_size: number;
-  id: string;
-  owner: string; // data item signer's public key
-  owner_address: string; // normalized address
-  signature: string;
-  tags: { name: string; value: string }[];
-  target: string;
-  anchor: string;
-}
